#130 Split unit tests from integration tests for NYISO client.
